content/canvas/test/test_2d.type.delete.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.type.delete</title>
<!-- Testing: window.CanvasRenderingContext2D is DontDelete -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

delete window.CanvasRenderingContext2D;
ok(window.CanvasRenderingContext2D !== undefined, "window.CanvasRenderingContext2D !== undefined");

SimpleTest.finish();

});
</script>
